users_1.1/requirements.txt
fastapi[standard]
sqlmodel
uvicorn
passlib
pyjwt